@@ -1,6 +1,6 @@
 {
   "name": "serverless-domain-manager",
-  "version": "5.1.2",
+  "version": "5.1.3",
   "engines": {
     "node": ">=12"
   },
@@ -44,27 +44,27 @@
     ]
   },
   "devDependencies": {
-    "@types/mocha": "^8.0.3",
-    "@types/node": "^14.11.1",
-    "aws-sdk-mock": "^5.1.0",
-    "chai": "^4.2.0",
+    "@types/mocha": "^8.2.3",
+    "@types/node": "^16.3.3",
+    "aws-sdk-mock": "^5.2.1",
+    "chai": "^4.3.4",
     "chai-spies": "^1.0.0",
-    "mocha": "^8.1.3",
+    "mocha": "^9.0.2",
     "mocha-param": "^2.0.1",
     "nyc": "^15.1.0",
-    "randomstring": "^1.1.5",
+    "randomstring": "^1.2.1",
     "request": "^2.88.2",
     "request-promise-native": "^1.0.9",
-    "serverless": "^2.1.1",
-    "serverless-plugin-split-stacks": "^1.9.3",
+    "serverless": "^2.52.0",
+    "serverless-plugin-split-stacks": "^1.11.2",
     "shelljs": "^0.8.4",
-    "ts-node": "^9.0.0",
+    "ts-node": "^10.1.0",
     "tslint": "^6.1.3",
-    "typescript": "^4.0.3"
+    "typescript": "^4.3.5"
   },
   "dependencies": {
-    "aws-sdk": "^2.756.0",
-    "chalk": "^4.1.0"
+    "aws-sdk": "^2.948.0",
+    "chalk": "^4.1.1"
   },
   "peerDependencies": {
     "serverless": "1 || 2"
