@@ -3,8 +3,8 @@
  */
 
 import {CloudFormation} from "aws-sdk";
-import DomainConfig = require("../DomainConfig");
-import Globals from "../Globals";
+import DomainConfig = require("../domain-config");
+import Globals from "../globals";
 import {getAWSPagedResults, throttledCall} from "../utils";
 
 class CloudFormationWrapper {
@@ -98,12 +98,12 @@ class CloudFormationWrapper {
         );
 
         // filter stacks by given stackName and check by nested stack RootId
+        const regex = new RegExp(`\/${stackName}\/`);
         const filteredStackNames = stacks
           .reduce((acc, stack) => {
               if (!stack.RootId) {
                   return acc;
               }
-              const regex = new RegExp(`\/${stackName}\/`);
               const match = stack.RootId.match(regex);
               if (match) {
                   acc.push(stack.StackName);
