@@ -2,8 +2,8 @@
 
 import APIGatewayWrapper = require("./aws/api-gateway-wrapper");
 import CloudFormationWrapper = require("./aws/cloud-formation-wrapper");
-import DomainConfig = require("./DomainConfig");
-import Globals from "./Globals";
+import DomainConfig = require("./domain-config");
+import Globals from "./globals";
 import {CustomDomain, ServerlessInstance, ServerlessOptions} from "./types";
 import {getAWSPagedResults, sleep, throttledCall} from "./utils";
 
